refactor(rebuild): move ratchet helpers out of hotspot
